// frontend/src/app/api/generated/model/notificationResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { NotificationType } from './notificationType';
import { UserReference } from './userReference';
import { NotificationTicketReference } from './notificationTicketReference';

export interface NotificationResponse {
  actor: UserReference | null;
  created_at: string;
  id: number;
  is_read: boolean;
  read_at: string | null;
  ticket: NotificationTicketReference | null;
  type: NotificationType;
}