// frontend/src/app/api/generated/model/notificationType.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export enum NotificationType {
  TICKET_ASSIGNED = 'TICKET_ASSIGNED',

  TICKET_REASSIGNED = 'TICKET_REASSIGNED',

  TICKET_PUBLIC_COMMENT = 'TICKET_PUBLIC_COMMENT',

  TICKET_IN_PROGRESS = 'TICKET_IN_PROGRESS',

  TICKET_RESOLVED = 'TICKET_RESOLVED',

  TICKET_CLOSED = 'TICKET_CLOSED',

  PASSWORD_RESET_REQUESTED = 'PASSWORD_RESET_REQUESTED',

  PASSWORD_RESET_COMPLETED = 'PASSWORD_RESET_COMPLETED',
}